tr1/shell: refactor
