--- src/tr1/specific/s_shell.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-void S_Shell_Init(void);
-void S_Shell_CreateWindow(void);
-void S_Shell_ToggleFullscreen(void);
-void S_Shell_HandleWindowResize(void);
