+++ src/tr2/game/shell.h
@@ -1,6 +1,3 @@
 #pragma once
 
-#include "game/shell/common.h"
-#include "game/shell/input.h"
-
 #include <libtrx/game/shell.h>
