--- src/tr2/game/shell/common.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-
-#include <libtrx/game/shell.h>
-
-void Shell_Start();
-
-void Shell_ProcessEvents(void);
-
-bool Shell_IsFullscreen(void);
