--- src/tr2/game/shell/input.h
@@ -1,1 +1,0 @@
-#pragma once
