+++ src/tr2/game/viewport.c
@@ -2,11 +2,11 @@
 
 #include "game/output.h"
 #include "game/render/common.h"
-#include "game/shell/common.h"
 #include "global/vars.h"
 
 #include <libtrx/config.h>
 #include <libtrx/game/math.h>
+#include <libtrx/game/shell.h>
 
 #define MAP_GAME_VARS()                                                        \
     MAP_GAME_VAR(win_max_x, g_PhdWinMaxX);                                     \
